some generator fixes
