--- lib/sidecar/grpc/service_resolver.ex
@@ -1,5 +1,0 @@
-defmodule Sidecar.GRPC.ServiceResolver do
-  @moduledoc """
-  This module is redefined at runtime during application start.
-  """
-end
