+++ spawn_sdk/spawn_sdk_example/lib/spawn_sdk_example/clock.pb.ex
