+++ spawn_sdk/spawn_sdk_example/priv/protos/actors/example.proto
