+++ spawn_sdk/spawn_sdk_example/priv/protos/actors/inside_folder/nested/clock.proto
