+++ spawn_sdk/spawn_sdk_example/priv/protos/modules/actors/inside_folder/nested/clock.pb.ex
@@ -1,158 +1,23 @@
-defmodule Io.Eigr.Spawn.Example.MyState do
+defmodule Io.Eigr.Spawn.Example.ClockActor.Service do
   @moduledoc false
-  use Protobuf, protoc_gen_elixir_version: "0.12.0", syntax: :proto3
-
-  def descriptor do
-    # credo:disable-for-next-line
-    %Google.Protobuf.DescriptorProto{
-      name: "MyState",
-      field: [
-        %Google.Protobuf.FieldDescriptorProto{
-          name: "value",
-          extendee: nil,
-          number: 1,
-          label: :LABEL_OPTIONAL,
-          type: :TYPE_INT32,
-          type_name: nil,
-          default_value: nil,
-          options: nil,
-          oneof_index: nil,
-          json_name: "value",
-          proto3_optional: nil,
-          __unknown_fields__: []
-        }
-      ],
-      nested_type: [],
-      enum_type: [],
-      extension_range: [],
-      extension: [],
-      options: nil,
-      oneof_decl: [],
-      reserved_range: [],
-      reserved_name: [],
-      __unknown_fields__: []
-    }
-  end
-
-  field :value, 1, type: :int32
-end
-
-defmodule Io.Eigr.Spawn.Example.MyBusinessMessage do
-  @moduledoc false
-  use Protobuf, protoc_gen_elixir_version: "0.12.0", syntax: :proto3
-
-  def descriptor do
-    # credo:disable-for-next-line
-    %Google.Protobuf.DescriptorProto{
-      name: "MyBusinessMessage",
-      field: [
-        %Google.Protobuf.FieldDescriptorProto{
-          name: "value",
-          extendee: nil,
-          number: 1,
-          label: :LABEL_OPTIONAL,
-          type: :TYPE_INT32,
-          type_name: nil,
-          default_value: nil,
-          options: nil,
-          oneof_index: nil,
-          json_name: "value",
-          proto3_optional: nil,
-          __unknown_fields__: []
-        }
-      ],
-      nested_type: [],
-      enum_type: [],
-      extension_range: [],
-      extension: [],
-      options: nil,
-      oneof_decl: [],
-      reserved_range: [],
-      reserved_name: [],
-      __unknown_fields__: []
-    }
-  end
-
-  field :value, 1, type: :int32
-end
-
-defmodule Io.Eigr.Spawn.Example.Joe.Service do
-  @moduledoc false
-  use GRPC.Service, name: "io.eigr.spawn.example.Joe", protoc_gen_elixir_version: "0.12.0"
+  use GRPC.Service, name: "io.eigr.spawn.example.ClockActor", protoc_gen_elixir_version: "0.12.0"
 
   def descriptor do
     # credo:disable-for-next-line
     %Google.Protobuf.FileDescriptorProto{
-      name: "example.proto",
+      name: "actors/inside_folder/nested/clock.proto",
       package: "io.eigr.spawn.example",
       dependency: ["google/api/annotations.proto", "google/protobuf/empty.proto"],
-      message_type: [
-        %Google.Protobuf.DescriptorProto{
-          name: "MyState",
-          field: [
-            %Google.Protobuf.FieldDescriptorProto{
-              name: "value",
-              extendee: nil,
-              number: 1,
-              label: :LABEL_OPTIONAL,
-              type: :TYPE_INT32,
-              type_name: nil,
-              default_value: nil,
-              options: nil,
-              oneof_index: nil,
-              json_name: "value",
-              proto3_optional: nil,
-              __unknown_fields__: []
-            }
-          ],
-          nested_type: [],
-          enum_type: [],
-          extension_range: [],
-          extension: [],
-          options: nil,
-          oneof_decl: [],
-          reserved_range: [],
-          reserved_name: [],
-          __unknown_fields__: []
-        },
-        %Google.Protobuf.DescriptorProto{
-          name: "MyBusinessMessage",
-          field: [
-            %Google.Protobuf.FieldDescriptorProto{
-              name: "value",
-              extendee: nil,
-              number: 1,
-              label: :LABEL_OPTIONAL,
-              type: :TYPE_INT32,
-              type_name: nil,
-              default_value: nil,
-              options: nil,
-              oneof_index: nil,
-              json_name: "value",
-              proto3_optional: nil,
-              __unknown_fields__: []
-            }
-          ],
-          nested_type: [],
-          enum_type: [],
-          extension_range: [],
-          extension: [],
-          options: nil,
-          oneof_decl: [],
-          reserved_range: [],
-          reserved_name: [],
-          __unknown_fields__: []
-        }
-      ],
+      message_type: [],
       enum_type: [],
       service: [
         %Google.Protobuf.ServiceDescriptorProto{
-          name: "Joe",
+          name: "ClockActor",
           method: [
             %Google.Protobuf.MethodDescriptorProto{
-              name: "Ping",
+              name: "Clock",
               input_type: ".google.protobuf.Empty",
-              output_type: ".io.eigr.spawn.example.MyBusinessMessage",
+              output_type: ".google.protobuf.Empty",
               options: %Google.Protobuf.MethodOptions{
                 deprecated: false,
                 idempotency_level: :IDEMPOTENCY_UNKNOWN,
@@ -164,7 +29,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
                     body: "",
                     additional_bindings: [],
                     response_body: "",
-                    pattern: {:get, "/v1/ping"},
+                    pattern: {:get, "/v1/clock"},
                     __unknown_fields__: []
                   }
                 },
@@ -175,9 +40,9 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
               __unknown_fields__: []
             },
             %Google.Protobuf.MethodDescriptorProto{
-              name: "Sum",
-              input_type: ".io.eigr.spawn.example.MyBusinessMessage",
-              output_type: ".io.eigr.spawn.example.MyBusinessMessage",
+              name: "SecondClock",
+              input_type: ".google.protobuf.Empty",
+              output_type: ".google.protobuf.Empty",
               options: %Google.Protobuf.MethodOptions{
                 deprecated: false,
                 idempotency_level: :IDEMPOTENCY_UNKNOWN,
@@ -189,7 +54,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
                     body: "*",
                     additional_bindings: [],
                     response_body: "",
-                    pattern: {:post, "/v1/sum"},
+                    pattern: {:post, "/v1/second_clock"},
                     __unknown_fields__: []
                   }
                 },
@@ -234,7 +99,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
         location: [
           %Google.Protobuf.SourceCodeInfo.Location{
             path: [],
-            span: [0, 0, 28, 1],
+            span: [0, 0, 24, 1],
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -321,104 +186,8 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
             __unknown_fields__: []
           },
           %Google.Protobuf.SourceCodeInfo.Location{
-            path: [4, 0],
-            span: [11, 0, 36],
-            leading_comments: nil,
-            trailing_comments: nil,
-            leading_detached_comments: [],
-            __unknown_fields__: []
-          },
-          %Google.Protobuf.SourceCodeInfo.Location{
-            path: [4, 0, 1],
-            span: [11, 8, 15],
-            leading_comments: nil,
-            trailing_comments: nil,
-            leading_detached_comments: [],
-            __unknown_fields__: []
-          },
-          %Google.Protobuf.SourceCodeInfo.Location{
-            path: [4, 0, 2, 0],
-            span: [11, 18, 34],
-            leading_comments: nil,
-            trailing_comments: nil,
-            leading_detached_comments: [],
-            __unknown_fields__: []
-          },
-          %Google.Protobuf.SourceCodeInfo.Location{
-            path: [4, 0, 2, 0, 5],
-            span: [11, 18, 23],
-            leading_comments: nil,
-            trailing_comments: nil,
-            leading_detached_comments: [],
-            __unknown_fields__: []
-          },
-          %Google.Protobuf.SourceCodeInfo.Location{
-            path: [4, 0, 2, 0, 1],
-            span: [11, 24, 29],
-            leading_comments: nil,
-            trailing_comments: nil,
-            leading_detached_comments: [],
-            __unknown_fields__: []
-          },
-          %Google.Protobuf.SourceCodeInfo.Location{
-            path: [4, 0, 2, 0, 3],
-            span: ~c"\v !",
-            leading_comments: nil,
-            trailing_comments: nil,
-            leading_detached_comments: [],
-            __unknown_fields__: []
-          },
-          %Google.Protobuf.SourceCodeInfo.Location{
-            path: [4, 1],
-            span: [13, 0, 46],
-            leading_comments: nil,
-            trailing_comments: nil,
-            leading_detached_comments: [],
-            __unknown_fields__: []
-          },
-          %Google.Protobuf.SourceCodeInfo.Location{
-            path: [4, 1, 1],
-            span: [13, 8, 25],
-            leading_comments: nil,
-            trailing_comments: nil,
-            leading_detached_comments: [],
-            __unknown_fields__: []
-          },
-          %Google.Protobuf.SourceCodeInfo.Location{
-            path: [4, 1, 2, 0],
-            span: [13, 28, 44],
-            leading_comments: nil,
-            trailing_comments: nil,
-            leading_detached_comments: [],
-            __unknown_fields__: []
-          },
-          %Google.Protobuf.SourceCodeInfo.Location{
-            path: [4, 1, 2, 0, 5],
-            span: [13, 28, 33],
-            leading_comments: nil,
-            trailing_comments: nil,
-            leading_detached_comments: [],
-            __unknown_fields__: []
-          },
-          %Google.Protobuf.SourceCodeInfo.Location{
-            path: [4, 1, 2, 0, 1],
-            span: ~c"\r\"'",
-            leading_comments: nil,
-            trailing_comments: nil,
-            leading_detached_comments: [],
-            __unknown_fields__: []
-          },
-          %Google.Protobuf.SourceCodeInfo.Location{
-            path: [4, 1, 2, 0, 3],
-            span: ~c"\r*+",
-            leading_comments: nil,
-            trailing_comments: nil,
-            leading_detached_comments: [],
-            __unknown_fields__: []
-          },
-          %Google.Protobuf.SourceCodeInfo.Location{
             path: [6, 0],
-            span: [15, 0, 28, 1],
+            span: [11, 0, 24, 1],
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -426,7 +195,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
           },
           %Google.Protobuf.SourceCodeInfo.Location{
             path: [6, 0, 1],
-            span: [15, 8, 11],
+            span: [11, 8, 18],
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -434,7 +203,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
           },
           %Google.Protobuf.SourceCodeInfo.Location{
             path: [6, 0, 2, 0],
-            span: [16, 2, 20, 3],
+            span: [12, 2, 16, 3],
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -442,7 +211,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
           },
           %Google.Protobuf.SourceCodeInfo.Location{
             path: [6, 0, 2, 0, 1],
-            span: [16, 6, 10],
+            span: [12, 6, 11],
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -450,7 +219,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
           },
           %Google.Protobuf.SourceCodeInfo.Location{
             path: [6, 0, 2, 0, 2],
-            span: [16, 11, 32],
+            span: ~c"\f\f!",
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -458,7 +227,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
           },
           %Google.Protobuf.SourceCodeInfo.Location{
             path: [6, 0, 2, 0, 3],
-            span: [16, 43, 60],
+            span: ~c"\f,A",
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -466,7 +235,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
           },
           %Google.Protobuf.SourceCodeInfo.Location{
             path: [6, 0, 2, 0, 4],
-            span: [17, 4, 19, 6],
+            span: [13, 4, 15, 6],
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -474,7 +243,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
           },
           %Google.Protobuf.SourceCodeInfo.Location{
             path: [6, 0, 2, 0, 4, 72_295_728],
-            span: [17, 4, 19, 6],
+            span: [13, 4, 15, 6],
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -482,7 +251,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
           },
           %Google.Protobuf.SourceCodeInfo.Location{
             path: [6, 0, 2, 1],
-            span: [22, 2, 27, 3],
+            span: [18, 2, 23, 3],
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -490,7 +259,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
           },
           %Google.Protobuf.SourceCodeInfo.Location{
             path: [6, 0, 2, 1, 1],
-            span: [22, 6, 9],
+            span: [18, 6, 17],
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -498,7 +267,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
           },
           %Google.Protobuf.SourceCodeInfo.Location{
             path: [6, 0, 2, 1, 2],
-            span: [22, 10, 27],
+            span: [18, 18, 39],
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -506,7 +275,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
           },
           %Google.Protobuf.SourceCodeInfo.Location{
             path: [6, 0, 2, 1, 3],
-            span: [22, 38, 55],
+            span: [18, 50, 71],
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -514,7 +283,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
           },
           %Google.Protobuf.SourceCodeInfo.Location{
             path: [6, 0, 2, 1, 4],
-            span: [23, 4, 26, 6],
+            span: [19, 4, 22, 6],
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -522,7 +291,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
           },
           %Google.Protobuf.SourceCodeInfo.Location{
             path: [6, 0, 2, 1, 4, 72_295_728],
-            span: [23, 4, 26, 6],
+            span: [19, 4, 22, 6],
             leading_comments: nil,
             trailing_comments: nil,
             leading_detached_comments: [],
@@ -539,7 +308,7 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
     }
   end
 
-  rpc(:Ping, Google.Protobuf.Empty, Io.Eigr.Spawn.Example.MyBusinessMessage, %{
+  rpc(:Clock, Google.Protobuf.Empty, Google.Protobuf.Empty, %{
     http: %{
       type: Google.Api.PbExtension,
       value: %Google.Api.HttpRule{
@@ -547,13 +316,13 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
         body: "",
         additional_bindings: [],
         response_body: "",
-        pattern: {:get, "/v1/ping"},
+        pattern: {:get, "/v1/clock"},
         __unknown_fields__: []
       }
     }
   })
 
-  rpc(:Sum, Io.Eigr.Spawn.Example.MyBusinessMessage, Io.Eigr.Spawn.Example.MyBusinessMessage, %{
+  rpc(:SecondClock, Google.Protobuf.Empty, Google.Protobuf.Empty, %{
     http: %{
       type: Google.Api.PbExtension,
       value: %Google.Api.HttpRule{
@@ -561,44 +330,43 @@ defmodule Io.Eigr.Spawn.Example.Joe.Service do
         body: "*",
         additional_bindings: [],
         response_body: "",
-        pattern: {:post, "/v1/sum"},
+        pattern: {:post, "/v1/second_clock"},
         __unknown_fields__: []
       }
     }
   })
 end
 
-defmodule Io.Eigr.Spawn.Example.Joe.ActorDispatcher do
+defmodule Io.Eigr.Spawn.Example.ClockActor.ActorDispatcher do
   @moduledoc since: "1.2.1"
-  use GRPC.Server, service: Io.Eigr.Spawn.Example.Joe.Service, http_transcode: true
+  use GRPC.Server, service: Io.Eigr.Spawn.Example.ClockActor.Service, http_transcode: true
 
   alias Sidecar.GRPC.Dispatcher
 
-  @spec ping(Google.Protobuf.Empty.t(), GRPC.Server.Stream.t()) ::
-          Io.Eigr.Spawn.Example.MyBusinessMessage.t()
-  def ping(message, stream) do
+  @spec clock(Google.Protobuf.Empty.t(), GRPC.Server.Stream.t()) :: Google.Protobuf.Empty.t()
+  def clock(message, stream) do
     request = %{
       system: "spawn-system",
-      actor_name: "Joe",
-      action_name: "Ping",
+      actor_name: "ClockActor",
+      action_name: "Clock",
       input: message,
       stream: stream,
-      descriptor: Io.Eigr.Spawn.Example.Joe.Service.descriptor()
+      descriptor: Io.Eigr.Spawn.Example.ClockActor.Service.descriptor()
     }
 
     Dispatcher.dispatch(request)
   end
 
-  @spec sum(Io.Eigr.Spawn.Example.MyBusinessMessage.t(), GRPC.Server.Stream.t()) ::
-          Io.Eigr.Spawn.Example.MyBusinessMessage.t()
-  def sum(message, stream) do
+  @spec second_clock(Google.Protobuf.Empty.t(), GRPC.Server.Stream.t()) ::
+          Google.Protobuf.Empty.t()
+  def second_clock(message, stream) do
     request = %{
       system: "spawn-system",
-      actor_name: "Joe",
-      action_name: "Sum",
+      actor_name: "ClockActor",
+      action_name: "SecondClock",
       input: message,
       stream: stream,
-      descriptor: Io.Eigr.Spawn.Example.Joe.Service.descriptor()
+      descriptor: Io.Eigr.Spawn.Example.ClockActor.Service.descriptor()
     }
 
     Dispatcher.dispatch(request)
@@ -612,6 +380,7 @@ defmodule Sidecar.GRPC.ProxyEndpoint do
   intercept(GRPC.Server.Interceptors.Logger)
 
   services = [
+    Io.Eigr.Spawn.Example.ClockActor.ActorDispatcher,
     Io.Eigr.Spawn.Example.Joe.ActorDispatcher
   ]
 
@@ -625,40 +394,6 @@ defmodule Sidecar.GRPC.ProxyEndpoint do
   run(services)
 end
 
-defmodule Sidecar.GRPC.ServiceResolver do
-  @moduledoc since: "1.2.1"
-
-  @actors [
-    {
-      "Joe",
-      %{
-        service_name: "Io.Eigr.Spawn.Example.Joe",
-        service_module: Io.Eigr.Spawn.Example.Joe.Service
-      }
-    }
-  ]
-
-  def has_actor?(actor_name) do
-    Enum.any?(@actors, fn {name, _} -> actor_name == name end)
-  end
-
-  def get_descriptor(actor_name) do
-    actor_attributes =
-      Enum.filter(@actors, fn {name, _} -> actor_name == name end)
-      |> Enum.map(fn {_name, attributes} -> attributes end)
-      |> List.first()
-
-    mod = Map.get(actor_attributes, :service_module)
-
-    mod.descriptor()
-    |> Map.get(:service)
-    |> Enum.filter(fn %Google.Protobuf.ServiceDescriptorProto{name: name} ->
-      actor_name == name
-    end)
-    |> List.first()
-  end
-end
-
 defmodule Sidecar.GRPC.Reflection.Server do
   @moduledoc since: "1.2.1"
 
@@ -666,6 +401,7 @@ defmodule Sidecar.GRPC.Reflection.Server do
     use GrpcReflection.Server,
       version: :v1,
       services: [
+        Io.Eigr.Spawn.Example.ClockActor.Service,
         Io.Eigr.Spawn.Example.Joe.Service
       ]
   end
@@ -674,6 +410,7 @@ defmodule Sidecar.GRPC.Reflection.Server do
     use GrpcReflection.Server,
       version: :v1alpha,
       services: [
+        Io.Eigr.Spawn.Example.ClockActor.Service,
         Io.Eigr.Spawn.Example.Joe.Service
       ]
   end
